Fix meta tag parsing for templates
