+++ src/Sulu/Bundle/AdminBundle/Metadata/FormMetadata/Loader/TemplateXmlLoader.php
@@ -13,6 +13,7 @@ namespace Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Loader;
 
 use Sulu\Bundle\AdminBundle\Exception\InvalidRootTagException;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\FormMetadata;
+use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\MetaXmlParser;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\PropertiesXmlParser;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\SchemaXmlParser;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\TagXmlParser;
@@ -42,6 +43,7 @@ class TemplateXmlLoader extends AbstractLoader
         private PropertiesXmlParser $propertiesXmlParser,
         private SchemaXmlParser $schemaXmlParser,
         private TagXmlParser $tagXmlParser,
+        private MetaXmlParser $metaXmlParser,
         private SchemaMetadataProvider $schemaMetadataProvider,
     ) {
         parent::__construct(
@@ -58,15 +60,23 @@ class TemplateXmlLoader extends AbstractLoader
 
         $form = new FormMetadata();
         $form->addResource($resource);
-        $formKey = $this->getValueFromXPath('/x:template/x:key', $xpath);
-        \assert(\is_string($formKey), 'Expected the template key of "' . $resource . '" to be defined.');
-        $form->setKey($formKey);
+        $templateKey = $this->getValueFromXPath('/x:template/x:key', $xpath);
+        \assert(\is_string($templateKey), 'Expected the template key of "' . $resource . '" to be defined.');
+        $form->setKey($templateKey);
 
         $tagNodes = $xpath->query('/x:template/x:tag') ?: [];
         $form->setTags($this->tagXmlParser->load($xpath, $tagNodes));
 
+        $templateNode = ($xpath->query('/x:template') ?: null)?->item(0);
+        \assert(null !== $templateNode, 'Expected <template> be defined for "' . $resource . '".');
+        $meta = $this->metaXmlParser->load($xpath, $templateNode);
+
+        if (\array_key_exists('title', $meta)) {
+            $form->setTitles($meta['title']);
+        }
+
         $propertiesNode = ($xpath->query('/x:template/x:properties') ?: null)?->item(0);
-        \assert(null !== $propertiesNode, 'Expected properties be defined for "' . $resource . '".');
+        \assert(null !== $propertiesNode, 'Expected <properties> be defined for "' . $resource . '".');
         $properties = $this->propertiesXmlParser->load(
             $xpath,
             $propertiesNode,
