+++ src/Sulu/Bundle/AdminBundle/Resources/config/services.xml
@@ -70,6 +70,12 @@
                  class="Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\PropertiesXmlParser"
                  public="false">
             <argument type="service" id="sulu_admin.tag_xml_parser" />
+            <argument type="service" id="sulu_admin.meta_xml_parser" />
+        </service>
+
+        <service id="sulu_admin.meta_xml_parser"
+                 class="Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\MetaXmlParser"
+                 public="false">
             <argument type="service" id="translator" />
             <argument>%sulu_core.translated_locales%</argument>
         </service>
@@ -175,6 +181,7 @@
             <argument type="service" id="sulu_admin.properties_xml_parser"/>
             <argument type="service" id="sulu_admin.schema_xml_parser"/>
             <argument type="service" id="sulu_admin.tag_xml_parser"/>
+            <argument type="service" id="sulu_admin.meta_xml_parser"/>
             <argument type="service" id="sulu_admin.schema_metadata_provider"/>
         </service>
