+++ src/Sulu/Bundle/AdminBundle/Tests/Functional/Controller/fixtures/overview.json
@@ -162,7 +162,7 @@
             "required": false,
             "multilingual": true,
             "spaceAfter": null,
-            "minOccurs": null,
+            "minOccurs": 0,
             "maxOccurs": null,
             "onInvalid": null,
             "tags": []
@@ -220,7 +220,7 @@
             "required": false,
             "multilingual": true,
             "spaceAfter": null,
-            "minOccurs": null,
+            "minOccurs": 0,
             "maxOccurs": null,
             "onInvalid": null,
             "tags": []
